IDP: fix crypt container sizing
It's imperative that when sizing the contents of partitions which will
hold encrypted containers, the aligned (up) sizes of the encapsulated
child partitions are used in the calculation. For the partitioned LUKS
scheme, this is critical to avoid the case where a partition may be
created too small for its intended image.

Misc:
Print the sizes of all partitions being created.
When aligning, catch overflow and make the calculation intent more obvious.

Ref #2